ai-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
openai>=1.14.3
pytesseract>=0.3.10
Pillow>=10.4.0
pdf2image>=1.17.0